@@ -1,63 +1,49 @@
 {
-  "name": "report",
+  "name": "client-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:ssr": "ng run Report:server:dev",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
-    "@ng-bootstrap/ng-bootstrap": "^5.3.1",
-    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
-    "bootstrap": "^4.6.0",
-    "core-js": "^3.8.3",
-    "jquery": "^3.5.1",
-    "node-sass": "^5.0.0",
-    "oidc-client": "^1.11.3",
-    "popper.js": "^1.16.0",
-    "protractor": "~5.4.2",
-    "rxjs": "^6.6.3",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0",
-    "zone.js": "0.9.1"
+    "@angular/animations": "~11.2.12",
+    "@angular/common": "~11.2.12",
+    "@angular/compiler": "~11.2.12",
+    "@angular/core": "~11.2.12",
+    "@angular/forms": "~11.2.12",
+    "@angular/localize": "~11.2.12",
+    "@angular/platform-browser": "~11.2.12",
+    "@angular/platform-browser-dynamic": "~11.2.12",
+    "@angular/router": "~11.2.12",
+    "@ng-bootstrap/ng-bootstrap": "^9.1.0",
+    "bootstrap": "^4.5.0",
+    "eslint-plugin-import": "^2.22.1",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@types/jasmine": "~3.4.4",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.2",
-    "ini": "^1.3.7",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
+    "@angular-devkit/build-angular": "~0.1102.11",
+    "@angular/cli": "~11.2.11",
+    "@angular/compiler-cli": "~11.2.12",
+    "@types/jasmine": "~3.6.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.1.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
-    "typescript": "3.5.3"
-  },
-  "optionalDependencies": {
-    "node-sass": "^5.0.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0"
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.1.5"
   }
 }
